requirements.txt
transformers==4.51.3
torch>=2.6.0
torchaudio==2.6.0
python-dotenv==1.1.0
numpy==2.2.5
accelerate==1.6.0
safetensors==0.5.3
huggingface_hub==0.30.2
librosa==0.11.0
tqdm==4.67.1
openai==1.75.0
rich==13.7.1
